+++ src/extensions/tableName.ts
@@ -1,14 +1,13 @@
 import * as z4 from "zod/v4/core";
+import { getZodMetadata, setZodMetadata } from "../utils/zodMetadataRegistry.js";
 
 export function tableName<T extends z4.$ZodObject>(this: T, name: string): T {
-    const currentMetadata = z4.globalRegistry.get(this);
+    const currentMetadata = getZodMetadata(this);
 
-    if (currentMetadata) {
-        currentMetadata.tableName = name;
-        z4.globalRegistry.add(this, currentMetadata);
-    } else {
-        z4.globalRegistry.add(this, { tableName: name });
-    }
+    setZodMetadata(this, {
+        ...currentMetadata,
+        tableName: name,
+    });
 
     return this;
 }