+++ src/parser/isForeignKey.ts
@@ -1,5 +1,6 @@
 import * as z4 from "zod/v4/core";
+import { getZodMetadata } from "../utils/zodMetadataRegistry.js";
 
 export function isForeignKey<T extends z4.$ZodType>(schema: T) {
-  return z4.globalRegistry.get(schema)?.foreignKey !== undefined;
+  return getZodMetadata(schema)?.foreignKey !== undefined;
 }